pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-tools"
version = "0.1.0"
dependencies = [
    "opencv-python-headless",
    "pillow",
    "requests",
    "pydantic>2.6",
    "tenacity"
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "twine",
    "ruff"
]

[tool.setuptools.packages.find]
include = ["vlm_tools*"]
exclude = [
    "notebooks",
    "examples",
]

[pytest]
log_cli = true
log_cli_level = "INFO"

[tool.setuptools.package-data]
"*" = ["*.csv"]

[tool.black]
line-length = 119
target-version = ['py38']
exclude = '''
    /(
        \.eggs
      | \.git
      | \.hg
      | \.mypy_cache
      | \.tox
      | \.venv
      | _build
      | buck-out
      | build
      | dist
    )/
'''

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # E501: Line too long
    "C901",  # C901: Function is too complex
    "B008",  # do not perform function calls in argument defaults
    "B904",  # B904: check for raise from None
]
line-length = 119
target-version = "py38"

[tool.pytest.ini_options]
addopts = "-sv -m 'not (skip)'"
